r/DailyProgrammer #228 Easy Letters in Alphabetical Order
https://www.reddit.com/r/dailyprogrammer/comments/3h9pde/20150817_challenge_228_easy_letters_in/

Working on this. I've got the list being iterated, but the list.sort()
function isn't working like I want. Produces:

('billowy', 'NOT IN ORDER')
('begins', 'NOT IN ORDER')
('bijoux', 'NOT IN ORDER')
('billowy', 'NOT IN ORDER')
('biopsy', 'NOT IN ORDER')
('chimps', 'NOT IN ORDER')
('chinos', 'NOT IN ORDER')
('chintz', 'NOT IN ORDER')
('defaced', 'NOT IN ORDER')
('fiddle', 'NOT IN ORDER')
('sponged', 'NOT IN ORDER')
('wronged', 'NOT IN ORDER')