There it is.
Needed to use "".join(sorted(item)) to sort the item into a list and
join those list items together. So: defaced becomes ['a', 'c', 'd', 'd',
'e', 'e', 'f'], the join makes it acddeef, and defaced != acddeef.

In the elif, reverse=True simply reverses defaced to decafed, and in
this instance is not equal.

Also, reminded of the importance of True and False being capitalized,
otherwise they are parsed (is that the correct word?) as variables
instead of boolean.

And, concatenation. I was printing as such print(item, "IN ORDER) which
would produce ('billowy', 'IN ORDER'). In order to get the output to
match the challenges output, with a hint from a comment post, I found
out that print(item + " IN ORDER") produces billowy IN ORDER without the
extra characters.